Add checkstyle with Gradle and fix stylechecks with the project.
