+++ config/checkstyle/suppressions.xml
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0"?>
+
+<!DOCTYPE suppressions PUBLIC
+		"-//Checkstyle//DTD SuppressionFilter Configuration 1.2//EN"
+		"https://checkstyle.org/dtds/suppressions_1_2.dtd">
+
+<suppressions>
+	<suppress checks="JavadocType" files=".*Test\.java"/>
+	<suppress checks="MissingJavadocMethodCheck" files=".*Test\.java"/>
+</suppressions>
