--- src/main/java/Pix/exception/PixException.java
@@ -1,12 +1,0 @@
-package Pix.exception;
-
-/**
- * The General Exception class that signals any errors.
- */
-public class PixException extends Exception {
-    public PixException() { }
-
-    public PixException(String errorMessage) {
-        super(errorMessage);
-    }
-}
