--- src/main/java/Pix/exception/PixUnknownCommandException.java
@@ -1,11 +1,0 @@
-package Pix.exception;
-
-/**
- * Triggers when the user inputs an unknown Pix.command.
- */
-public class PixUnknownCommandException extends PixException {
-    @Override
-    public String getMessage() {
-        return "Unknown Command: Do you even know what you're typing..?";
-    }
-}
