+++ src/main/java/pix/exception/PixInvalidTaskException.java
@@ -1,0 +1,11 @@
+package pix.exception;
+
+/**
+ * Triggers when the format of the specified Pix.task is wrong.
+ */
+public class PixInvalidTaskException extends PixException {
+    @Override
+    public String getMessage() {
+        return "The format to enter the specified Pix.task is wrong";
+    }
+}
