+++ src/main/java/pix/exception/PixUnknownCommandException.java
@@ -1,0 +1,11 @@
+package pix.exception;
+
+/**
+ * Triggers when the user inputs an unknown Pix.command.
+ */
+public class PixUnknownCommandException extends PixException {
+    @Override
+    public String getMessage() {
+        return "Unknown Command: Do you even know what you're typing..?";
+    }
+}
